Move compile only structs from abi.rs to tests.rs
