+++ src/evm/tests.rs
@@ -1,10 +1,934 @@
-use super::abi::{ArbAggregator, ArbGasInfo, ArbInfo, ArbOwner, ArbReplayableTx, ArbStatistics};
 use super::*;
 use crate::compile::miniconstants::init_constant_table;
-use crate::run::{load_from_file, RuntimeEnvironment};
+use crate::run::{load_from_file, Machine, RuntimeEnvironment};
 use crate::uint256::Uint256;
-use ethers_signers::Signer;
+use ethers_signers::{Signer, Wallet};
 use std::path::Path;
+
+pub struct ArbInfo {
+    pub contract_abi: AbiForContract,
+    debug: bool,
+}
+
+impl ArbInfo {
+    pub fn _new(debug: bool) -> Self {
+        let mut contract_abi =
+            AbiForContract::new_from_file(&builtin_contract_path("ArbInfo")).unwrap();
+        contract_abi.bind_interface_to_address(Uint256::from_u64(101));
+        ArbInfo {
+            contract_abi,
+            debug,
+        }
+    }
+
+    pub fn _get_balance(
+        &self,
+        machine: &mut Machine,
+        addr: &Uint256,
+    ) -> Result<Uint256, ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            Uint256::from_u64(1112),
+            "getBalance",
+            &[ethabi::Token::Address(addr.to_h160())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            let return_vals = ethabi::decode(
+                &[ethabi::ParamType::Uint(256)],
+                &receipts[0].get_return_data(),
+            )?;
+            match return_vals[0] {
+                ethabi::Token::Uint(ui) => Ok(Uint256::from_u256(&ui)),
+                _ => panic!(),
+            }
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+}
+
+pub struct ArbOwner<'a> {
+    pub contract_abi: AbiForContract,
+    wallet: &'a Wallet,
+    my_address: Uint256,
+    debug: bool,
+}
+
+impl<'a> ArbOwner<'a> {
+    pub fn _new(wallet: &'a Wallet, debug: bool) -> Self {
+        let mut contract_abi =
+            AbiForContract::new_from_file(&builtin_contract_path("ArbOwner")).unwrap();
+        contract_abi.bind_interface_to_address(Uint256::from_u64(107));
+        ArbOwner {
+            contract_abi,
+            wallet,
+            my_address: Uint256::from_bytes(wallet.address().as_bytes()),
+            debug,
+        }
+    }
+
+    pub fn _give_ownership(
+        &self,
+        machine: &mut Machine,
+        new_owner: Uint256,
+        old_owner: Option<Uint256>,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            old_owner.unwrap_or(self.my_address.clone()),
+            "giveOwnership",
+            &[ethabi::Token::Address(new_owner.to_h160())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from(format!(
+                "tx failed: {}",
+                receipts[0]._get_return_code_text()
+            )))
+        }
+    }
+
+    pub fn _add_to_reserve_funds(
+        &self,
+        machine: &mut Machine,
+        amount: Uint256,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "addToReserveFunds",
+            &[],
+            machine,
+            amount,
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _set_fees_enabled(
+        &self,
+        machine: &mut Machine,
+        enabled: bool,
+        force_owner: bool, // force the message to come from address zero, which is an owner
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            if force_owner {
+                Uint256::zero()
+            } else {
+                self.my_address.clone()
+            },
+            "setFeesEnabled",
+            &[ethabi::Token::Bool(enabled)],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from(format!(
+                "tx failed: {}",
+                receipts[0]._get_return_code_text()
+            )))
+        }
+    }
+
+    pub fn _get_fee_recipients(
+        &self,
+        machine: &mut Machine,
+    ) -> Result<(Uint256, Uint256), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "getFeeRecipient",
+            &[],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if !receipts[0].succeeded() {
+            return Err(ethabi::Error::from("reverted"));
+        }
+
+        let return_vals = ethabi::decode(
+            &[ethabi::ParamType::Address, ethabi::ParamType::Address],
+            &receipts[0].get_return_data(),
+        )?;
+
+        match (&return_vals[0], &return_vals[1]) {
+            (ethabi::Token::Address(addr1), ethabi::Token::Address(addr2)) => Ok((
+                Uint256::from_bytes(addr1.as_bytes()),
+                Uint256::from_bytes(addr2.as_bytes()),
+            )),
+            _ => panic!(),
+        }
+    }
+
+    pub fn _set_fee_recipients(
+        &self,
+        machine: &mut Machine,
+        recipient1: Uint256,
+        recipient2: Uint256,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "setFeeRecipient",
+            &[
+                ethabi::Token::Address(recipient1.to_h160()),
+                ethabi::Token::Address(recipient2.to_h160()),
+            ],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _set_seconds_per_send(
+        &self,
+        machine: &mut Machine,
+        seconds_per_send: Uint256,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "setSecondsPerSend",
+            &[ethabi::Token::Uint(seconds_per_send.to_u256())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _set_gas_accounting_params(
+        &self,
+        machine: &mut Machine,
+        speed_limit: Uint256,
+        gas_pool_max: Uint256,
+        tx_gas_limit: Uint256,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "setGasAccountingParams",
+            &[
+                ethabi::Token::Uint(speed_limit.to_u256()),
+                ethabi::Token::Uint(gas_pool_max.to_u256()),
+                ethabi::Token::Uint(tx_gas_limit.to_u256()),
+            ],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _change_sequencer(
+        &self,
+        machine: &mut Machine,
+        sequencer_addr: Uint256,
+        delay_seconds: Uint256,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "changeSequencer",
+            &[
+                ethabi::Token::Address(sequencer_addr.to_h160()),
+                ethabi::Token::Uint(delay_seconds.to_u256()),
+            ],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _start_code_upload(&self, machine: &mut Machine) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function_compressed(
+            self.my_address.clone(),
+            "startCodeUpload",
+            &[],
+            machine,
+            Uint256::zero(),
+            self.wallet,
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _continue_code_upload(
+        &self,
+        machine: &mut Machine,
+        marshalled_code: Vec<u8>,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "continueCodeUpload",
+            &[ethabi::Token::Bytes(marshalled_code)],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _finish_code_upload_as_arbos_upgrade(
+        &self,
+        machine: &mut Machine,
+    ) -> Result<(), ethabi::Error> {
+        let _res = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "finishCodeUploadAsArbosUpgrade",
+            &[],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+        Ok(())
+    }
+
+    pub fn _finish_code_upload_as_pluggable(
+        &self,
+        machine: &mut Machine,
+        id: Uint256,
+        keep_state: bool,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "finishCodeUploadAsPluggable",
+            &[
+                ethabi::Token::Uint(id.to_u256()),
+                ethabi::Token::Bool(keep_state),
+            ],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+}
+
+pub struct ArbGasInfo<'a> {
+    pub contract_abi: AbiForContract,
+    _wallet: &'a Wallet,
+    my_address: Uint256,
+    debug: bool,
+}
+
+impl<'a> ArbGasInfo<'a> {
+    pub fn _new(wallet: &'a Wallet, debug: bool) -> Self {
+        let mut contract_abi =
+            AbiForContract::new_from_file(&builtin_contract_path("ArbGasInfo")).unwrap();
+        contract_abi.bind_interface_to_address(Uint256::from_u64(108));
+        ArbGasInfo {
+            contract_abi,
+            _wallet: wallet,
+            my_address: Uint256::from_bytes(wallet.address().as_bytes()),
+            debug,
+        }
+    }
+
+    pub fn _get_prices_in_wei(
+        &self,
+        machine: &mut Machine,
+    ) -> Result<(Uint256, Uint256, Uint256, Uint256, Uint256, Uint256), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "getPricesInWei",
+            &[],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            let return_vals = ethabi::decode(
+                &[
+                    ethabi::ParamType::Uint(256),
+                    ethabi::ParamType::Uint(256),
+                    ethabi::ParamType::Uint(256),
+                    ethabi::ParamType::Uint(256),
+                    ethabi::ParamType::Uint(256),
+                    ethabi::ParamType::Uint(256),
+                ],
+                &receipts[0].get_return_data(),
+            )?;
+
+            match (
+                &return_vals[0],
+                &return_vals[1],
+                &return_vals[2],
+                &return_vals[3],
+                &return_vals[4],
+                &return_vals[5],
+            ) {
+                (
+                    ethabi::Token::Uint(ui0),
+                    ethabi::Token::Uint(ui1),
+                    ethabi::Token::Uint(ui2),
+                    ethabi::Token::Uint(ui3),
+                    ethabi::Token::Uint(ui4),
+                    ethabi::Token::Uint(ui5),
+                ) => Ok((
+                    Uint256::from_u256(&ui0),
+                    Uint256::from_u256(&ui1),
+                    Uint256::from_u256(&ui2),
+                    Uint256::from_u256(&ui3),
+                    Uint256::from_u256(&ui4),
+                    Uint256::from_u256(&ui5),
+                )),
+                _ => panic!(),
+            }
+        } else {
+            Err(ethabi::Error::from(format!(
+                "tx failed: {}",
+                receipts[0]._get_return_code_text()
+            )))
+        }
+    }
+
+    pub fn _get_prices_in_arbgas(
+        &self,
+        machine: &mut Machine,
+    ) -> Result<(Uint256, Uint256, Uint256), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "getPricesInArbGas",
+            &[],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            let return_vals = ethabi::decode(
+                &[
+                    ethabi::ParamType::Uint(256),
+                    ethabi::ParamType::Uint(256),
+                    ethabi::ParamType::Uint(256),
+                ],
+                &receipts[0].get_return_data(),
+            )?;
+
+            match (&return_vals[0], &return_vals[1], &return_vals[2]) {
+                (ethabi::Token::Uint(ui0), ethabi::Token::Uint(ui1), ethabi::Token::Uint(ui2)) => {
+                    Ok((
+                        Uint256::from_u256(&ui0),
+                        Uint256::from_u256(&ui1),
+                        Uint256::from_u256(&ui2),
+                    ))
+                }
+                _ => panic!(),
+            }
+        } else {
+            Err(ethabi::Error::from(format!(
+                "tx failed: {}",
+                receipts[0]._get_return_code_text()
+            )))
+        }
+    }
+
+    pub fn _get_gas_accounting_params(
+        &self,
+        machine: &mut Machine,
+    ) -> Result<(Uint256, Uint256, Uint256), ethabi::Error> {
+        let (receipts, _sends) = self.contract_abi.call_function(
+            self.my_address.clone(),
+            "getGasAccountingParams",
+            &[],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if receipts.len() != 1 {
+            return Err(ethabi::Error::from("wrong number of receipts"));
+        }
+
+        if receipts[0].succeeded() {
+            let return_vals = ethabi::decode(
+                &[
+                    ethabi::ParamType::Uint(256),
+                    ethabi::ParamType::Uint(256),
+                    ethabi::ParamType::Uint(256),
+                ],
+                &receipts[0].get_return_data(),
+            )?;
+
+            match (&return_vals[0], &return_vals[1], &return_vals[2]) {
+                (ethabi::Token::Uint(ui0), ethabi::Token::Uint(ui1), ethabi::Token::Uint(ui2)) => {
+                    Ok((
+                        Uint256::from_u256(&ui0),
+                        Uint256::from_u256(&ui1),
+                        Uint256::from_u256(&ui2),
+                    ))
+                }
+                _ => panic!(),
+            }
+        } else {
+            Err(ethabi::Error::from(format!(
+                "tx failed: {}",
+                receipts[0]._get_return_code_text()
+            )))
+        }
+    }
+}
+
+pub struct ArbAggregator {
+    pub contract_abi: AbiForContract,
+    debug: bool,
+}
+
+impl ArbAggregator {
+    pub fn _new(debug: bool) -> Self {
+        let mut contract_abi =
+            AbiForContract::new_from_file(&builtin_contract_path("ArbAggregator")).unwrap();
+        contract_abi.bind_interface_to_address(Uint256::from_u64(109));
+        ArbAggregator {
+            contract_abi,
+            debug,
+        }
+    }
+
+    pub fn _get_preferred_aggregator(
+        &self,
+        machine: &mut Machine,
+        addr: Uint256,
+    ) -> Result<(Uint256, bool), ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            Uint256::zero(), // send from address zero
+            "getPreferredAggregator",
+            &[ethabi::Token::Address(addr.to_h160())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            let return_vals = ethabi::decode(
+                &[ethabi::ParamType::Address, ethabi::ParamType::Bool],
+                &receipts[0].get_return_data(),
+            )?;
+
+            match (&return_vals[0], &return_vals[1]) {
+                (ethabi::Token::Address(r1), ethabi::Token::Bool(r2)) => {
+                    Ok((Uint256::from_bytes(r1.as_bytes()), *r2))
+                }
+                _ => panic!(),
+            }
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _set_preferred_aggregator(
+        &self,
+        machine: &mut Machine,
+        addr: Uint256,
+        sender: Uint256,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            sender,
+            "setPreferredAggregator",
+            &[ethabi::Token::Address(addr.to_h160())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _get_default_aggregator(&self, machine: &mut Machine) -> Result<Uint256, ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            Uint256::zero(), // send from address zero
+            "getDefaultAggregator",
+            &[],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            let return_vals = ethabi::decode(
+                &[ethabi::ParamType::Address],
+                &receipts[0].get_return_data(),
+            )?;
+
+            match &return_vals[0] {
+                ethabi::Token::Address(r1) => Ok(Uint256::from_bytes(r1.as_bytes())),
+                _ => panic!(),
+            }
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _set_default_aggregator(
+        &self,
+        machine: &mut Machine,
+        addr: Uint256,
+        sender: Option<Uint256>, // default sender is address zero
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            if let Some(s) = sender {
+                s
+            } else {
+                Uint256::zero()
+            },
+            "setDefaultAggregator",
+            &[ethabi::Token::Address(addr.to_h160())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+}
+
+pub struct ArbReplayableTx {
+    pub contract_abi: AbiForContract,
+    debug: bool,
+}
+
+impl ArbReplayableTx {
+    pub fn _new(debug: bool) -> Self {
+        let mut contract_abi =
+            AbiForContract::new_from_file(&builtin_contract_path("ArbRetryableTx")).unwrap();
+        contract_abi.bind_interface_to_address(Uint256::from_u64(110));
+        ArbReplayableTx {
+            contract_abi,
+            debug,
+        }
+    }
+
+    pub fn _redeem(&self, machine: &mut Machine, txid: Uint256) -> Result<(), ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            Uint256::zero(), // send from address zero
+            "redeem",
+            &[ethabi::Token::FixedBytes(txid.to_bytes_be())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() < 1) || (receipts.len() > 2) || (sends.len() != 0) {
+            println!("{} receipts, {} sends", receipts.len(), sends.len());
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[receipts.len() - 1].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _get_timeout(
+        &self,
+        machine: &mut Machine,
+        txid: Uint256,
+    ) -> Result<Uint256, ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            Uint256::zero(), // send from address zero
+            "getTimeout",
+            &[ethabi::Token::FixedBytes(txid.to_bytes_be())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            Ok(Uint256::from_bytes(&receipts[0].get_return_data()))
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _get_lifetime(&self, machine: &mut Machine) -> Result<Uint256, ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            Uint256::zero(), // send from address zero
+            "getLifetime",
+            &[],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            Ok(Uint256::from_bytes(&receipts[0].get_return_data()))
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _get_keepalive_price(
+        &self,
+        machine: &mut Machine,
+        txid: Uint256,
+    ) -> Result<(Uint256, Uint256), ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            Uint256::zero(), // send from address zero
+            "getKeepalivePrice",
+            &[ethabi::Token::FixedBytes(txid.to_bytes_be())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            let return_data = receipts[0].get_return_data();
+            Ok((
+                Uint256::from_bytes(&return_data[0..32]),
+                Uint256::from_bytes(&return_data[32..64]),
+            ))
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _keepalive(
+        &self,
+        machine: &mut Machine,
+        txid: Uint256,
+        payment: Uint256,
+    ) -> Result<Uint256, ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            Uint256::zero(), // send from address zero
+            "keepalive",
+            &[ethabi::Token::FixedBytes(txid.to_bytes_be())],
+            machine,
+            payment,
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            Ok(Uint256::from_bytes(&receipts[0].get_return_data()))
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _get_beneficiary(
+        &self,
+        machine: &mut Machine,
+        txid: Uint256,
+    ) -> Result<Uint256, ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            Uint256::zero(), // send from address zero
+            "getBeneficiary",
+            &[ethabi::Token::FixedBytes(txid.to_bytes_be())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            Ok(Uint256::from_bytes(&receipts[0].get_return_data()))
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+
+    pub fn _cancel(
+        &self,
+        machine: &mut Machine,
+        txid: Uint256,
+        sender: Uint256,
+    ) -> Result<(), ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            sender,
+            "cancel",
+            &[ethabi::Token::FixedBytes(txid.to_bytes_be())],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            Ok(())
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+}
+
+pub struct ArbStatistics {
+    pub contract_abi: AbiForContract,
+    debug: bool,
+}
+
+impl ArbStatistics {
+    pub fn _new(debug: bool) -> Self {
+        let mut contract_abi =
+            AbiForContract::new_from_file(&builtin_contract_path("ArbStatistics")).unwrap();
+        contract_abi.bind_interface_to_address(Uint256::from_u64(111));
+        ArbStatistics {
+            contract_abi,
+            debug,
+        }
+    }
+
+    pub fn _get_stats(
+        &self,
+        machine: &mut Machine,
+    ) -> Result<(Uint256, Uint256, Uint256, Uint256, Uint256, Uint256), ethabi::Error> {
+        let (receipts, sends) = self.contract_abi.call_function(
+            Uint256::zero(), // send from address zero
+            "getStats",
+            &[],
+            machine,
+            Uint256::zero(),
+            self.debug,
+        )?;
+
+        if (receipts.len() != 1) || (sends.len() != 0) {
+            println!("{} receipts, {} sends", receipts.len(), sends.len());
+            Err(ethabi::Error::from("wrong number of receipts or sends"))
+        } else if receipts[0].succeeded() {
+            let retdata = receipts[0].get_return_data();
+            Ok((
+                Uint256::from_bytes(&retdata[0..32]),
+                Uint256::from_bytes(&retdata[32..64]),
+                Uint256::from_bytes(&retdata[64..96]),
+                Uint256::from_bytes(&retdata[96..128]),
+                Uint256::from_bytes(&retdata[128..160]),
+                Uint256::from_bytes(&retdata[160..192]),
+            ))
+        } else {
+            Err(ethabi::Error::from("reverted"))
+        }
+    }
+}
 
 #[test]
 fn test_payment_to_self() {
